requirements.txt
streamlit
pillow
streamlit-drawable-canvas
pandas